refactor: consolidate Flying Broom rendering system
- Create shared FlyingBroomRenderHelper with common rendering utilities
- Add IFlyingBroomRenderer interface to enforce consistent method signatures
- Refactor both entity and item renderers to use shared components
- Remove duplicate model files (entity models now use item variants)
- Move renderers to dedicated renderers/ package structure
- Eliminate ~300 lines of duplicated rendering code
- Fix entity renderer overlay texture issue
