+++ src/main/java/org/sosly/witchcraft/entities/EntityTypeRegistry.java
@@ -9,6 +9,7 @@ import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
 import org.sosly.witchcraft.Witchcraft;
+import org.sosly.witchcraft.entities.tools.FlyingBroomEntity;
 
 @Mod.EventBusSubscriber(modid = Witchcraft.MOD_ID, bus = Mod.EventBusSubscriber.Bus.MOD)
 public class EntityTypeRegistry {
