+++ src/main/java/org/sosly/witchcraft/items/FlyingBroomItem.java
@@ -7,12 +7,9 @@ import net.minecraft.commands.arguments.EntityAnchorArgument;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.network.chat.Component;
-import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
-import net.minecraft.world.InteractionResultHolder;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.DyeColor;
-import net.minecraft.world.item.DyeItem;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
@@ -20,7 +17,6 @@ import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.block.Block;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.LayeredCauldronBlock;
 import net.minecraft.world.phys.Vec3;
 import net.minecraftforge.client.extensions.common.IClientItemExtensions;
@@ -30,8 +26,8 @@ import org.jetbrains.annotations.NotNull;
 import org.sosly.witchcraft.api.capabilities.ICovenCapability;
 import org.sosly.witchcraft.capabilities.coven.CovenProvider;
 import org.sosly.witchcraft.entities.EntityTypeRegistry;
-import org.sosly.witchcraft.entities.FlyingBroomEntity;
-import org.sosly.witchcraft.items.renderers.FlyingBroomItemRenderer;
+import org.sosly.witchcraft.entities.tools.FlyingBroomEntity;
+import org.sosly.witchcraft.renderers.items.FlyingBroomItemRenderer;
 
 import java.util.List;
 import java.util.function.Consumer;
