--- src/main/resources/assets/mnaw/models/entity/flying_broom_brush.json
@@ -1,14 +1,0 @@
-{
-  "parent": "forge:item/default",
-  "loader": "forge:composite",
-  "children": {
-    "brush": {
-      "loader": "forge:obj",
-      "automatic_culling": true,
-      "shade_quads": true,
-      "flip_v": true,
-      "emissive_ambient": false,
-      "model": "mnaw:models/obj/broom_brush.obj"
-    }
-  }
-}
