--- src/main/resources/assets/mnaw/models/entity/flying_broom_connectors.json
@@ -1,8 +1,0 @@
-{
-  "loader": "forge:obj",
-  "automatic_culling": true,
-  "shade_quads": true,
-  "flip_v": true,
-  "emissive_ambient": false,
-  "model": "mnaw:models/obj/broom_connectors.obj"
-}
